Bug in page frame allocator, working with virtual addresses messes things up

+++ src/util/dbgprinter.h
@@ -4,5 +4,5 @@
 
 void dbg_print(const char * str);
 void panic(const char * str);
-char* itoa(int, int);
+char* itoa(int64_t, int);
 #endif
